+++ src/library.rs
@@ -236,7 +236,7 @@ pub struct BaseConfig {
     /// The path to where the database file should be stored,
     /// e.g. `/home/foo/.local/share/bliss-rs/bliss.db`
     pub database_path: PathBuf,
-    /// The latest features version a song has been analyzed
+    /// The latest features' version a song has been analyzed
     /// with.
     pub features_version: u16,
     /// The number of CPU cores an analysis will be performed with.
@@ -391,6 +391,17 @@ pub struct Library<Config, D: ?Sized> {
     decoder: PhantomData<D>,
 }
 
+/// Hold an error that happened while processing songs during analysis.
+#[derive(Debug, Eq, PartialEq)]
+pub struct ProcessingError {
+    /// The path of the song whose analysis was attempted.
+    pub song_path: PathBuf,
+    /// The actual error string.
+    pub error: String,
+    /// Features version the analysis was attempted with.
+    pub features_version: u16,
+}
+
 /// Struct holding both a Bliss song, as well as any extra info
 /// that a user would want to store in the database related to that
 /// song.
@@ -443,7 +454,7 @@ impl<Config: AppConfigTrait, D: ?Sized + DecoderTrait> Library<Config, D> {
                 cue_path text,
                 audio_file_path text,
                 stamp timestamp default current_timestamp,
-                version integer,
+                version integer not null,
                 analyzed boolean default false,
                 extra_info json,
                 error text
@@ -485,6 +496,39 @@ impl<Config: AppConfigTrait, D: ?Sized + DecoderTrait> Library<Config, D> {
                 foreign key(song_2_id) references song(id) on delete cascade,
                 foreign key(odd_one_out_id) references song(id) on delete cascade
             )
+        ",
+        // Add the "not null" constraint to the "version" column
+        "
+            create table song_bak (
+                id integer primary key,
+                path text not null unique,
+                duration float,
+                album_artist text,
+                artist text,
+                title text,
+                album text,
+                track_number integer,
+                disc_number integer,
+                genre text,
+                cue_path text,
+                audio_file_path text,
+                stamp timestamp default current_timestamp,
+                version integer not null,
+                analyzed boolean default false,
+                extra_info json,
+                error text
+            );
+            insert into song_bak (
+                id, path, duration, album_artist, artist, title, album, track_number,
+                disc_number,genre, cue_path, audio_file_path, stamp, version,
+                analyzed, extra_info, error
+            ) select
+                id, path, duration, album_artist, artist, title, album, track_number,
+                disc_number,genre, cue_path, audio_file_path, stamp,
+                coalesce(version, 1), analyzed, extra_info, error
+            from song;
+            drop table song;
+            alter table song_bak rename to song;
         ",
     ];
 
@@ -974,7 +1018,7 @@ impl<Config: AppConfigTrait, D: ?Sized + DecoderTrait> Library<Config, D> {
             return Ok(());
         }
         log::info!(
-            "Analyzing {} songs, this might take some time…",
+            "Analyzing {} song(s), this might take some time…",
             number_songs
         );
         let pb = if show_progress_bar {
@@ -1417,15 +1461,40 @@ impl<Config: AppConfigTrait, D: ?Sized + DecoderTrait> Library<Config, D> {
             .unwrap()
             .execute(
                 "
-            insert or replace into song (path, error) values (?1, ?2)
+            insert or replace into song (path, error, version) values (?1, ?2, ?3)
             ",
-                [
+                params![
                     song_path.into().to_string_lossy().to_string(),
                     e.to_string(),
+                    // At this point, FEATURES_VERSION is the best indicator we have
+                    // of the version (since we don't have a proper Song).
+                    FEATURES_VERSION,
                 ],
             )
             .map_err(|e| BlissError::ProviderError(e.to_string()))?;
         Ok(())
+    }
+
+    /// Return all the songs that failed the analysis.
+    pub fn get_failed_songs(&self) -> Result<Vec<ProcessingError>> {
+        let conn = self.sqlite_conn.lock().unwrap();
+        let mut stmt = conn.prepare(
+            "
+            select path, error, version
+                from song where error is not null order by id
+            ",
+        )?;
+        let rows = stmt.query_map([], |row| {
+            Ok(ProcessingError {
+                song_path: row.get::<_, String>(0)?.into(),
+                error: row.get(1)?,
+                features_version: row.get(2)?,
+            })
+        })?;
+        Ok(rows
+            .into_iter()
+            .map(|r| r.unwrap())
+            .collect::<Vec<ProcessingError>>())
     }
 
     /// Delete a song with path `song_path` from the database.
@@ -1838,52 +1907,52 @@ mod test {
                     insert into song (
                         id, path, artist, title, album, album_artist, track_number,
                         disc_number, genre, duration, analyzed, version, extra_info,
-                        cue_path, audio_file_path
+                        cue_path, audio_file_path, error
                     ) values (
                         1001, '/path/to/song1001', 'Artist1001', 'Title1001', 'An Album1001',
                         'An Album Artist1001', 3, 1, 'Electronica1001', 310, true,
                         1, '{\"ignore\": true, \"metadata_bliss_does_not_have\":
-                        \"/path/to/charlie1001\"}', null, null
+                        \"/path/to/charlie1001\"}', null, null, null
                     ),
                     (
                         2001, '/path/to/song2001', 'Artist2001', 'Title2001', 'An Album2001',
                         'An Album Artist2001', 2, 1, 'Electronica2001', 410, true,
                         1, '{\"ignore\": false, \"metadata_bliss_does_not_have\":
-                        \"/path/to/charlie2001\"}', null, null
+                        \"/path/to/charlie2001\"}', null, null, null
                     ),
                     (
                         2201, '/path/to/song2201', 'Artist2001', 'Title2001', 'An Album2001',
                         'An Album Artist2001', 1, 2, 'Electronica2001', 410, true,
                         1, '{\"ignore\": false, \"metadata_bliss_does_not_have\":
-                        \"/path/to/charlie2201\"}', null, null
+                        \"/path/to/charlie2201\"}', null, null, null
                     ),
                     (
                         3001, '/path/to/song3001', null, null, null,
-                        null, null, null, null, null, false, 1, '{}', null, null
+                        null, null, null, null, null, false, 1, '{}', null, null, null
                     ),
                     (
                         4001, '/path/to/song4001', 'Artist4001', 'Title4001', 'An Album4001',
                         'An Album Artist4001', 1, 1, 'Electronica4001', 510, true,
                         0, '{\"ignore\": false, \"metadata_bliss_does_not_have\":
-                        \"/path/to/charlie4001\"}', null, null
+                        \"/path/to/charlie4001\"}', null, null, null
                     ),
                     (
                         5001, '/path/to/song5001', 'Artist5001', 'Title5001', 'An Album1001',
                         'An Album Artist5001', 1, 1, 'Electronica5001', 610, true,
                         1, '{\"ignore\": false, \"metadata_bliss_does_not_have\":
-                        \"/path/to/charlie5001\"}', null, null
+                        \"/path/to/charlie5001\"}', null, null, null
                     ),
                     (
                         6001, '/path/to/song6001', 'Artist6001', 'Title6001', 'An Album2001',
                         'An Album Artist6001', 1, 1, 'Electronica6001', 710, true,
                         1, '{\"ignore\": false, \"metadata_bliss_does_not_have\":
-                        \"/path/to/charlie6001\"}', null, null
+                        \"/path/to/charlie6001\"}', null, null, null
                     ),
                     (
                         7001, '/path/to/song7001', 'Artist7001', 'Title7001', 'An Album7001',
                         'An Album Artist7001', 1, 1, 'Electronica7001', 810, true,
                         1, '{\"ignore\": false, \"metadata_bliss_does_not_have\":
-                        \"/path/to/charlie7001\"}', null, null
+                        \"/path/to/charlie7001\"}', null, null, null
                     ),
                     (
                         7002, '/path/to/cuetrack.cue/CUE_TRACK001', 'CUE Artist',
@@ -1891,7 +1960,7 @@ mod test {
                         'CUE Album Artist', 1, 1, null, 810, true,
                         1, '{\"ignore\": false, \"metadata_bliss_does_not_have\":
                         \"/path/to/charlie7001\"}', '/path/to/cuetrack.cue',
-                        '/path/to/cuetrack.flac'
+                        '/path/to/cuetrack.flac', null
                     ),
                     (
                         7003, '/path/to/cuetrack.cue/CUE_TRACK002', 'CUE Artist',
@@ -1899,19 +1968,29 @@ mod test {
                         'CUE Album Artist', 2, 1, null, 910, true,
                         1, '{\"ignore\": false, \"metadata_bliss_does_not_have\":
                         \"/path/to/charlie7001\"}', '/path/to/cuetrack.cue',
-                        '/path/to/cuetrack.flac'
+                        '/path/to/cuetrack.flac', null
                     ),
                     (
                         8001, '/path/to/song8001', 'Artist8001', 'Title8001', 'An Album1001',
                         'An Album Artist8001', 3, 1, 'Electronica8001', 910, true,
                         0, '{\"ignore\": false, \"metadata_bliss_does_not_have\":
-                        \"/path/to/charlie8001\"}', null, null
+                        \"/path/to/charlie8001\"}', null, null, null
                     ),
                     (
                         9001, './data/s16_stereo_22_5kHz.flac', 'Artist9001', 'Title9001',
                         'An Album9001', 'An Album Artist8001', 3, 1, 'Electronica8001',
                         1010, true, 0, '{\"ignore\": false, \"metadata_bliss_does_not_have\":
-                        \"/path/to/charlie7001\"}', null, null
+                        \"/path/to/charlie7001\"}', null, null, null
+                    ),
+                    (
+                        404, './data/not-existing.m4a', null, null,
+                        null, null, null, null, null,
+                        null, false, 0, null, null, null, 'error finding the file'
+                    ),
+                    (
+                        502, './data/invalid-file.m4a', null, null,
+                        null, null, null, null, null,
+                        null, false, 0, null, null, null, 'error decoding the file'
                     );
                     ",
                     [],
@@ -3441,7 +3520,7 @@ mod test {
     fn test_library_new_database_upgrade() {
         let config_dir = TempDir::new("tmp").unwrap();
         let sqlite_db_path = config_dir.path().join("test.db");
-        // Initialize the database with the contents of old_database.sq, without
+        // Initialize the database with the contents of old_database.sql, without
         // having anything to do with Library (yet)
         {
             let sqlite_conn = Connection::open(sqlite_db_path.clone()).unwrap();
@@ -3486,7 +3565,7 @@ mod test {
         let version: u32 = sqlite_conn
             .query_row("pragma user_version", [], |row| row.get(0))
             .unwrap();
-        assert_eq!(version, 4);
+        assert_eq!(version, 5);
         // Make sure we can call this over and over without any problem
         Library::<BaseConfig, DummyDecoder>::new_from_base(
             Some(config_dir.path().join("config.txt")),
@@ -3497,7 +3576,7 @@ mod test {
         let version: u32 = sqlite_conn
             .query_row("pragma user_version", [], |row| row.get(0))
             .unwrap();
-        assert_eq!(version, 4);
+        assert_eq!(version, 5);
     }
 
     #[test]
@@ -3521,7 +3600,7 @@ mod test {
         let version: u32 = sqlite_conn
             .query_row("pragma user_version", [], |row| row.get(0))
             .unwrap();
-        assert_eq!(version, 4);
+        assert_eq!(version, 5);
     }
 
     #[test]
@@ -3811,5 +3890,27 @@ mod test {
         )
         .unwrap();
         assert!(config_dir.is_dir());
+    }
+
+    #[test]
+    #[cfg(feature = "ffmpeg")]
+    fn test_library_get_failed_songs() {
+        let (library, _temp_dir, _) = setup_test_library();
+        let failed_songs = library.get_failed_songs().unwrap();
+        assert_eq!(
+            failed_songs,
+            vec![
+                ProcessingError {
+                    song_path: PathBuf::from("./data/not-existing.m4a"),
+                    error: String::from("error finding the file"),
+                    features_version: 0,
+                },
+                ProcessingError {
+                    song_path: PathBuf::from("./data/invalid-file.m4a"),
+                    error: String::from("error decoding the file"),
+                    features_version: 0,
+                }
+            ]
+        );
     }
 }