+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## bliss 0.10.1
+* Add a way to retrieve failed analysis for a Library.
 * Update dependencies.
 
 ## bliss 0.10.0